% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MovingBubbles.R
\name{MovingBubbles}
\alias{MovingBubbles}
\title{Animated Bubble Chart}
\usage{
MovingBubbles(df, key, value, frame, color = NULL, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{df}{dataframe}

\item{key}{name of key column}

\item{value}{name of value column}

\item{frame}{name of frame column}

\item{color}{name of color column}
}
\description{
MovingBubbles can be used to plot bubble chart with transitions.
}
\examples{
dat <- data.frame(alphabets = rep(letters[1:6],5),
                 size = round(runif(30)*100),
                 time = rep(paste0(1:5, "pm"), each = 6))
MovingBubbles(dat,"alphabets", "size", "time")

}
