% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MovingBubbles.R
\name{MovingBubbles}
\alias{MovingBubbles}
\title{Animated Bubble Chart}
\usage{
MovingBubbles(df, key, frame, value, color = NULL, bubble_size = 1,
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{df}{dataframe}

\item{key}{name of key column}

\item{frame}{name of frame column}

\item{value}{name of value column}

\item{color}{dataframe with key and color columns}

\item{bubble_size}{bubble size factor}
}
\description{
MovingBubbles can be used to plot bubble chart with transitions.
}
\examples{
dat <- data.frame(data = rep(letters[1:6],5),
                 value = round(runif(30)*100),
                 time = rep(paste0(1:5, "pm"), each = 6),
                 value2 = runif(30))
colordf <- data.frame(key = unique(dat$data), color = terrain.colors(6))
MovingBubbles(dat, key = "data", frame = "time", value = "value")
MovingBubbles(dat, key = "data", frame = "time", value = "value", color = colordf)

}
